requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-admin-rangefilter==0.5.4
django-import-export==2.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-redis==4.11.0
djangorestframework==3.11.0
django-silk==4.0.1
et-xmlfile==1.0.1
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
